+++ src/view/inventario/ModalProducto/FormularioProdStyle.css
@@ -12,6 +12,15 @@
     box-shadow: inset 0px 4px 4px rgba(0, 0, 0, 0.25);
     border-radius: 36px;
   }
+
+  .MuiOutlinedInput-root {
+    border-radius: 15px !important;
+  }
+
+  .option {
+    
+    background-color: #65b1f367 !important;
+  }
   
   .foco {
     display: flex;
