+++ src/view/ventas/Ventas.js
@@ -1,4 +1,4 @@
-import FormularioProd from "../inventario/Componentes/Modal/FormularioProd.js"
+import FormularioProd from "../inventario/ModalProducto/FormularioProd.js"
 import { useState } from "react";
 
 const Ventas = () => {
@@ -6,14 +6,7 @@ const Ventas = () => {
   return (
     <>
       <h1>Hola Soy Ventas</h1>
-      <FormularioProd
-            recarga={recarga}
-            setRecarga={setRecarga}
-            tipo="Producto"
-            metodo="post"
-            titulo="Crear Producto"
-            
-          />
+      
     </>
   );
 };
